server/requirements.txt
openenv-core[core]>=0.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.40.0
pytest>=8.0.0
pydantic>=2.0.0
